elevaweb/requirements.txt
aldryn-apphooks-config==0.7.0
asgiref==3.7.2
astor==0.8.1
click==8.1.7
cssselect2==0.7.0
dj-database-url==2.1.0
Django==3.2
django-app-enabler==0.2.0
django-appdata==0.4.0
django-classy-tags==3.0.1
django-cms==3.9.0
django-filer==3.0.3
django-formtools==2.4
django-js-asset==2.0.0
django-meta==2.3.0
django-mptt==0.14.0
django-parler==2.3
django-polymorphic==3.1.0
django-sekizai==3.0.1
django-sortedm2m==3.1.1
django-taggit==4.0.0
django-taggit-autosuggest==0.4.1
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-treebeard==4.4
djangocms-admin-style==2.0.2
djangocms-apphook-setup==0.5.0
djangocms-attributes-field==3.0.0
djangocms-blog==2.0.1
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.1
djangocms-googlemap==2.1.0
djangocms-icon==2.0.0
djangocms-installer==2.0.0
djangocms-link==3.1.1
djangocms-picture==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
easy-thumbnails==2.8.5
html5lib==1.1
lxml==4.9.3
packaging==23.1
Pillow==10.0.0
PyMySQL==1.1.0
python-decouple==3.8
pytz==2023.3
reportlab==4.0.4
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.7.1
tzlocal==5.0.1
webencodings==0.5.1
